feature-comments -> Реализация комментариев

+++ main-service/src/main/java/ru/practicum/mainservice/service/CategoryServiceImpl.java
@@ -5,12 +5,13 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.practicum.mainservice.dto.category.CategoryDTO;
-import ru.practicum.mainservice.dto.category.CreateCategoryDTO;
+import ru.practicum.mainservice.dto.category.CategoryDto;
+import ru.practicum.mainservice.dto.category.CreateCategoryDto;
 import ru.practicum.mainservice.exception.APIException;
 import ru.practicum.mainservice.mapper.CategoryMapper;
 import ru.practicum.mainservice.model.Category;
 import ru.practicum.mainservice.repository.CategoryRepository;
+import ru.practicum.mainservice.repository.EventRepository;
 import ru.practicum.mainservice.util.OffsetBasedPageRequest;
 
 import java.util.List;
@@ -18,14 +19,15 @@ import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
-public class CategoryServiceImpl implements CategoryService {
+public class CategoryServiceImpl implements CategoryService { //Вызов идет не по репозиторию, а через маппер по причине рекурсии вызовов
 
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final EventRepository eventRepository;
 
     @Override
     @Transactional
-    public CategoryDTO create(CreateCategoryDTO category) {
+    public CategoryDto create(CreateCategoryDto category) {
         Category newCategory = new Category();
         newCategory.setName(category.getName());
         return categoryMapper.toDto(categoryRepository.save(newCategory));
@@ -33,7 +35,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional(readOnly = true)
-    public CategoryDTO getById(int categoryId) {
+    public CategoryDto getById(int categoryId) {
         Category category = getCategoryById(categoryId);
         return categoryMapper.toDto(category);
     }
@@ -52,12 +54,17 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     public void delete(int categoryId) {
         Category category = getCategoryById(categoryId);
-        categoryRepository.delete(category);
+        if (category != null) {
+            boolean hasEventsInCategory = eventRepository.findAll().contains(category);
+            if (!hasEventsInCategory) {
+                categoryRepository.delete(category);
+            }
+        }
     }
 
     @Override
     @Transactional
-    public CategoryDTO update(int categoryId, CreateCategoryDTO category) {
+    public CategoryDto update(int categoryId, CreateCategoryDto category) {
         Category fromDb = getCategoryById(categoryId);
         fromDb.setName(category.getName());
         return categoryMapper.toDto(fromDb);
@@ -65,7 +72,7 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     @Transactional(readOnly = true)
-    public List<CategoryDTO> getAll(int from, int size) {
+    public List<CategoryDto> getAll(int from, int size) {
         Pageable pageable = new OffsetBasedPageRequest(from, size);
         return categoryRepository.findAll(pageable).getContent().stream()
                 .map(categoryMapper::toDto).collect(Collectors.toList());
